refactor : Point를 Position으로 수정
