+++ src/test/kotlin/next/step/minesweeper/domain/position/PointTest.kt
@@ -1,4 +1,4 @@
-package next.step.minesweeper.domain.point
+package next.step.minesweeper.domain.position
 
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
@@ -11,17 +11,17 @@ class PointTest : DescribeSpec({
         context("생성") {
             it("위치 y가 0보다 작으면 예외 발생") {
                 assertThrows<IllegalArgumentException> {
-                    Point(0, -1)
+                    Position(0, -1)
                 }.shouldHaveMessage("위치 y는 0 보다 작을 수 없습니다.")
             }
 
             it("위치 x가 0보다 작으면 예외 발생") {
                 assertThrows<IllegalArgumentException> {
-                    Point(-1, 0)
+                    Position(-1, 0)
                 }.shouldHaveMessage("위치 x는 0 보다 작을 수 없습니다.")
             }
             it("base는 기준이 되는 포인트 제공") {
-                Point.base() shouldBe Point(0, 0)
+                Position.base() shouldBe Position(0, 0)
             }
         }
     }
